backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.5
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
pypdf==5.0.1
google-cloud-aiplatform==1.68.0
google-auth==2.35.0
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1

